build(test): unblock MVS pipeline for migrated tests
Two follow-ups to PR #48 surfaced during the first end-to-end
make clean build link run:

1. test/mvs/tst{arext,lstr,vpol}.c used <stdint.h> / uintptr_t,
   which c2asm370 (GCC 3.2.3) does not provide. Removed the header
   include and rewrote the two (uintptr_t) pointer-strip casts as
   (unsigned long), which is portable across 24-bit MVS and 64-bit
   Linux.

2. Dataset space for SOURCE and LOAD was undersized for the new
   test corpus:
   - SOURCE: 34 generated .s files total ~2.4 MB, tstbifs.s alone
     ~800 KB. Primary bumped from 50 to 200 TRK, secondary 20 to
     50 TRK, directory 30 to 50 blocks.
   - LOAD hit IEC032I E37-04 on IBMUSER.REXX370.V0R1M0D.LOAD during
     the 11th consecutive link (TSTHELO): the 16-extent PDS limit
     was exhausted because secondary extents were only 10 TRK. Primary
     bumped from 20 to 150 TRK, secondary 10 to 50 TRK, directory 10
     to 30 blocks.

After the reallocation, make build produces RC=0 for all 34 modules
and make link produces RC=0 for all 18 load modules (TSTANCH + 15
test modules + IRX#HELO + IRXJCL).

+++ test/mvs/tstarext.c
@@ -13,7 +13,6 @@
 /*  (c) 2026 mvslovers - REXX/370 Project                            */
 /* ------------------------------------------------------------------ */
 
-#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -55,7 +54,7 @@ static int tests_failed = 0;
 static Lstr make_lstr(const char *cstr)
 {
     Lstr s;
-    s.pstr = (unsigned char *)(uintptr_t)cstr;
+    s.pstr = (unsigned char *)(unsigned long)cstr;
     s.len = strlen(cstr);
     s.maxlen = s.len;
     s.type = LSTRING_TY;
